details page and delete car
